hydrate-timelines

+++ hydrate-timelines.py
@@ -20,8 +20,6 @@ class HydrateTimelines(object):
     def _user_files(self, user_id):
         file_names = glob.glob(
             os.path.join(self._user_path(user_id), '*.json.gz'))
-        # temporary hack
-        file_names = [f for f in file_names if '-07' not in f]
         return file_names
 
     def hydrate(self):
@@ -29,13 +27,15 @@ class HydrateTimelines(object):
             print('processing user {} #{}/{}...'.format(
                 user_id, i, len(self.user_ids)))
             for infile in self._user_files(user_id):
-                outfile = new_file_name(infile)
-                try:
-                    os.remove(outfile)
-                except OSError:
-                    pass
-                hydrate_file('key-and-secret.txt', 'auth.txt',
-                             infile, outfile, 'error.log')
+                # TODO: temporary hack
+                if '-07.json.gz' not in infile:
+                    outfile = new_file_name(infile)
+                    try:
+                        os.remove(outfile)
+                    except OSError:
+                        pass
+                    hydrate_file('key-and-secret.txt', 'auth.txt',
+                                 infile, outfile, 'error.log')
 
 
 if __name__ == '__main__':
